third_party/blink/web_tests/external/wpt/speculation-rules/prerender/referrer-policy-not-accepted.html

<!DOCTYPE html>
<meta name="timeout" content="long">
<title>Referrer policy must be sufficiently strict to allow prerendering</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/html/browsers/browsing-the-web/remote-context-helper/resources/remote-context-helper.js"></script>
<script src="resources/utils.js"></script>

<script>
setup(() => assertSpeculationRulesIsSupported());

promise_test(async t => {
  const {exec, tryToActivate} =
      await create_prerendered_page(t, {}, {referrer_policy: 'unsafe-url'});
  const result = await tryToActivate();
  assert_equals(result, 'discarded');
}, 'do not prerender with "unsafe-url"');

</script>
